requirements.txt
jsonrpclib-pelix==0.4.3.2
pytest==7.4.0
